ui: Fix bug where attack style gets reset to null
When the items get updated for the attack styles combo box, the state was always being overwritten with null. 
This was caused by the unnecessary function override StateBoundJComboBox::setItems which adds:
```
if (!Objects.equals(prev, getValue()))
{
			toState();
}
```
In the if statement, getValue() always evaluates to null because the above call to `super.setItems` clears the combobox entries and refills them from scratch, thus resetting the selectedIndex to 0. You can validate this by adding `super.setNullLast(true);` to the AttackStyleSelectPanel constructor and observing that the attackstyle always resets to the first one in the list instead of null